moved the getFullUserProfile to UserAccountService
