+++ services/ResumeService.ts
@@ -3,8 +3,6 @@ import { Service } from 'typedi';
 import { EntityManager } from 'typeorm';
 import { InjectManager } from 'typeorm-typedi-extensions';
 import { ResumeModel } from '../models/ResumeModel';
-import { UserModel } from '../models/UserModel';
-import { PrivateProfile } from '../types';
 import Repositories, { TransactionsManager } from '../repositories';
 
 @Service()
@@ -20,13 +18,5 @@ export default class ResumeService {
       .resume(txn).findVisibleResumes());
     if (!resumes) throw new NotFoundError('Resumes was not found');
     return resumes;
-  }
-
-  public async getFullUserProfile(user: UserModel) : Promise<PrivateProfile>{
-    return this.transactions.readOnly(async (txn) => {
-      const userProfile = user.getFullUserProfile();
-      userProfile.resumes = await Repositories.resume(txn).findAllByUser(user);
-      return userProfile;
-    });
   }
 }
